Подключил все компоненты страницы "Profile"
